# Conflicts:
#	GPOS Winter Project 2019/Assets/Scenes/BattleScene.unity
#	GPOS Winter Project 2019/Assets/Scripts/Core/WaveManager.cs
#	GPOS Winter Project 2019/Assets/Scripts/Projectile/Arrow.cs
#	GPOS Winter Project 2019/Assets/Scripts/Projectile/DeathBall.cs
#	GPOS Winter Project 2019/Assets/Scripts/Projectile/Explosive.cs
#	GPOS Winter Project 2019/Assets/Scripts/Unit/NPC/Dragon.cs
#	GPOS Winter Project 2019/Assets/Scripts/Unit/NPC/Goblin.cs
#	GPOS Winter Project 2019/Assets/Scripts/Unit/NPC/Lich.cs
#	GPOS Winter Project 2019/Assets/Scripts/Unit/NPC/NPC.cs
#	GPOS Winter Project 2019/Assets/Scripts/Unit/Player.cs
#	GPOS Winter Project 2019/Assets/Scripts/Unit/Unit.cs
#	GPOS Winter Project 2019/Assets/resources/Prefabs/Dragon.prefab

+++ GPOS Winter Project 2019/Assets/Scripts/Unit/NPC/Archer.cs
@@ -60,7 +60,7 @@ public class Archer : NPC , IMissileAttack
         if (Vector2.Distance(Target.position, this.position) <= ArcherMissileRange)
         {
             if (ArcherMissileCool > MissileCool) return;
-            GameObject.Find("ProjectileFactory").GetComponent<ProjectileFactoryManager>().PlaceProjectile("Arrow", this, this.position, Target.position, (int)ArcherAttack, 10f, 1f);
+            GameObject.Find("ProjectileFactory").GetComponent<ProjectileFactoryManager>().PlaceProjectile("Arrow", this, this.position, Target.position, ArcherAttack, 10f, 1f);
             MissileCool = 0;
         }
     }
